Reorganize preassembly tests.
